thermal: tsens: Update mdm9640 TSENS fuse map
Add support to distinguish fusemap between mdm9640 v1/v2 and
to read the EEPROM fuse data for calibration to calculate the
gain and offset that is used to compute the final temperature.

Update the DT document with target name reference for mdm9640
along with the addition for 9640v2 compatible string support.

CRs-Fixed: 1008416
Change-Id: I8dfbbc094c5d857dc904cf18cb462fc82c46dd57
Signed-off-by: Rama Krishna Phani A <rphani@codeaurora.org>